+++ changelog.md
@@ -6,10 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Added
 - Added .gitignore to exclude log files from version control.
 - Memory files 
+- Scripts for improving UI
 
 ### Changed
 - Modified nn.c to handle my data as vectors
 - Created better neuron structure in neuron.c and neuron_layer.c
+- output via dla and not just terminal
 
 ### Removed
 - Logs deleted